Move navigator_gpu to src/common/util
