+++ src/webgpu/api/validation/error_scope.spec.ts
@@ -17,8 +17,8 @@ TODO: consider slightly revising these tests to make sure they're complete. {
 
 import { Fixture } from '../../../common/framework/fixture.js';
 import { makeTestGroup } from '../../../common/framework/test_group.js';
+import { getGPU } from '../../../common/util/navigator_gpu.js';
 import { assert, raceWithRejectOnTimeout } from '../../../common/util/util.js';
-import { getGPU } from '../../util/navigator_gpu.js';
 
 class F extends Fixture {
   _device: GPUDevice | undefined = undefined;
