+++ src/webgpu/util/device_pool.ts
@@ -1,4 +1,5 @@
 import { SkipTestCase } from '../../common/framework/fixture.js';
+import { getGPU } from '../../common/util/navigator_gpu.js';
 import {
   assert,
   raceWithRejectOnTimeout,
@@ -6,8 +7,6 @@ import {
   assertReject,
 } from '../../common/util/util.js';
 import { DefaultLimits } from '../constants.js';
-
-import { getGPU } from './navigator_gpu.js';
 
 export interface DeviceProvider {
   acquire(): GPUDevice;
